Solved refresh problem!
I deleted the class and all of its methods.  Turns out, they were probably just fine.  Oh well.  The problem was that every time a button other than the "Process" button was pushed, it would drop the main df variable and have to start all over again because it registered that there was no document uploaded.  I got rid of the button and just made the whole analysis function run as soon as a file is uploaded.  Now I need to figure out why it is re-running the function every time even though I have cached it.  Pretty sure it is because it's an uploaded document instead of a pre-loaded dataset.  Ah, the joys of live data! :smile: